requirements.txt

streamlit
torch
transformers
pillow
pyngrok
